TS declaration declares currency global (#297)
* TS declaration declares currency global

ISSUE: The currency export is also declared globally as a namespace const and constructor.

When importing currency.js once in a module, TypeScript loads this declaration file and declares `currency` global and assumes it's available in all modules. This is not true and will give runtime errors. Exporting`currency` inside the module declaration ensures it's is imported in every module that is using it because TS will give compile errors.

* formatting

Co-authored-by: Jason <jason@scurker.com>